Fixed convert to sphinx for Wooster
Convert to sphinx now takes an additional argument, scan_pos, which specifies whether the subject was scanned in the HFS or HFP position. Default is set to HFP, which calls MRI_convert for a second time to flip the sphinx position. This could possibly be cleaned if MRI_convert will take both the --sphinx argument along with --in_orientation and --out_orientation.